Second version of the main listviewq.
The best way to use an array of strings as a model is with an "ArrayAdapter", but hopefully our model won't always be an array, so I've decided to use a "BaseAdapter", and to put it in a separate file. That, and because that way I can see the interface methods better. Also, we had to take the context with us and create a constructor for it on the adapter.