Creation of our first model "Photo".
In this commit, we are preparing to implement an SQLite database. We've started by creating a model for our Photo, with its basic constructor and only its getters (I figured setters made no sense, since we'll hopefully be using a database soon). The rest of the changes are just to adapt the project to our way of developing :-)