Modification of the adapter to inherit from "CursorAdapter".
We've changed the "PhotoAdapter" to inherit from "CursorAdapter" rather than "BaseAdapter", deleting the unnecessary methods and implementing the required ones ( "getView()" and "bindView()" ). We've also fixed a bug in our "PhotoDAO" class constructor that was causing our app to crash. Then, we've created another method to select a single Photo based on its ID. Finally, we've created a constructor in our "Photo" model based on a given cursor.